refactor: move 11 large inline test blocks to sibling tests.rs (#862)
Pure relocation of large inline #[cfg(test)] mod tests { ... } blocks
into sibling <mod>/tests.rs files per the migration trigger in
.claude/rules/03-unit-testing.md, mirroring the existing db/src/books.rs
+ db/src/books/tests.rs (and db/src/ebook/{accent,cover,stat}) pattern.
No test logic changed.

server/src/rate_limit.rs and frontend/src/pages/landing/sorting.rs are
intentionally excluded — PR #1059 (in flight) already touches test names
in those same inline blocks; a fast-follow should cover them once #1059
merges.
